Reorganize pciusb controller sources
