+++ rom/usb/pciusb/uhci/uhcichip.c
