+++ rom/usb/pciusb/uhci/uhcichip.h
