+++ rom/usb/pciusb/uhwcmd.c
@@ -11,10 +11,10 @@
 #include LC_LIBDEFS_FILE
 
 #include "uhwcmd.h"
-#include "ohciproto.h"
-#include "uhciproto.h"
-#include "ehciproto.h"
-#include "xhciproto.h"
+#include "ohci/ohciproto.h"
+#include "uhci/uhciproto.h"
+#include "ehci/ehciproto.h"
+#include "xhci/xhciproto.h"
 
 static inline BOOL uhwIsRootHubIOReq(const struct IOUsbHWReq *ioreq,
                                      const struct PCIUnit *unit)
